chore(metrics): rename `writing` feature to `record`
