+++ lez/sequencer/core/Cargo.toml
@@ -12,7 +12,7 @@ lee.workspace = true
 lee_core.workspace = true
 chain_state.workspace = true
 common.workspace = true
-sequencer_core_metrics = { workspace = true, features = ["writing"] }
+sequencer_core_metrics = { workspace = true, features = ["record"] }
 storage.workspace = true
 mempool.workspace = true
 logos-blockchain-zone-sdk.workspace = true
