+++ lez/sequencer/core/metrics/Cargo.toml
@@ -9,8 +9,8 @@ workspace = true
 
 [features]
 default = []
-# Enable metrics writing
-writing = ["dep:common", "dep:metrics", "dep:strum"]
+# Enable metrics record
+record = ["dep:common", "dep:metrics", "dep:strum"]
 
 [dependencies]
 common = { workspace = true, optional = true }
