+++ lez/sequencer/core/metrics/src/lib.rs
@@ -1,9 +1,9 @@
 //! This crate provides all metrics exposed by the sequencer core crate.
 
-#[cfg(feature = "writing")]
-pub use writing::*;
+#[cfg(feature = "record")]
+pub use record::*;
 
 pub mod names;
 
-#[cfg(feature = "writing")]
-pub mod writing;
+#[cfg(feature = "record")]
+pub mod record;
