+++ lez/sequencer/service/Cargo.toml
@@ -15,7 +15,7 @@ mempool.workspace = true
 sequencer_core = { workspace = true, features = ["testnet"] }
 sequencer_service_protocol.workspace = true
 sequencer_service_rpc = { workspace = true, features = ["server"] }
-sequencer_service_metrics = { workspace = true, features = ["writing"] }
+sequencer_service_metrics = { workspace = true, features = ["record"] }
 programs.workspace = true
 
 clap = { workspace = true, features = ["derive", "env"] }
